requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
exceptiongroup==1.2.0
Flask==3.0.1
Flask-PyMongo==2.3.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
pluggy==1.3.0
pymongo==4.6.1
pytest==7.4.4
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==3.0.1
